+++ tests/helper.php
@@ -29,18 +29,22 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class qtype_pmatchjme_test_helper extends question_test_helper {
+
+    #[\Override]
     public function get_test_questions() {
-        return array('benzene');
+        return ['benzene'];
     }
 
     /**
-     * @return qtype_pmatchjme_question
+     * Create a test question instance for drawing a benzene molecule.
+     *
+     * @return qtype_pmatchjme_question The prepared question object.
      */
     public function make_pmatchjme_question_benzene() {
         question_bank::load_question_definition_classes('pmatchjme');
         $question = new qtype_pmatchjme_question();
 
-        test_question_maker::initialise_a_question($dd);
+        test_question_maker::initialise_a_question($question);
 
         $question->name = 'Draw benzene';
         $question->questiontext = 'Please draw a benzene molecule.';
@@ -48,37 +52,42 @@ class qtype_pmatchjme_test_helper extends question_test_helper {
         $question->qtype = question_bank::get_qtype('pmatchjme');
         $question->allowsubscript = false;
         $question->allowsuperscript = false;
+        $question->quotematching = 0;
 
-        $question->answers = array(
+        $question->answers = [
             1 => new qtype_pmatchjme_answer(1, 'match (c1ccccc1)', 1, 'Well done!', FORMAT_HTML, 0),
             2 => new qtype_pmatchjme_answer(2, '*', 0, 'That is not right.', FORMAT_HTML, 1),
-        );
+        ];
 
         return $question;
     }
 
     /**
+     * Get form data for creating a benzene pmatchjme question.
+     *
      * @return stdClass data to create a pmatchjme question.
      */
     public function get_pmatchjme_question_form_data_benzene() {
         $fromform = new stdClass();
 
         $fromform->name = 'Draw benzene';
-        $fromform->questiontext = array('text' => 'Please draw a benzene molecule.', 'format' => FORMAT_HTML);
+        $fromform->questiontext = ['text' => 'Please draw a benzene molecule.', 'format' => FORMAT_HTML];
         $fromform->defaultmark = 1.0;
-        $fromform->generalfeedback = array('text' => 'This is the simplest aromatic molecule.', 'format' => FORMAT_HTML);
+        $fromform->generalfeedback = ['text' => 'This is the simplest aromatic molecule.', 'format' => FORMAT_HTML];
         $fromform->allowsubscript = 0;
         $fromform->allowsuperscript = 0;
-        $fromform->synonymsdata = array();
+        $fromform->modelanswer = 'c1ccccc1';
+        $fromform->quotematching = 0;
+        $fromform->synonymsdata = [];
 
-        $fromform->answer = array('match (c1ccccc1)');
-        $fromform->fraction = array('1');
-        $fromform->feedback = array(
-                array('text' => 'Well done!', 'format' => FORMAT_HTML),
-        );
-        $fromform->atomcount = array('0');
+        $fromform->answer = ['match (c1ccccc1)'];
+        $fromform->fraction = ['1'];
+        $fromform->feedback = [
+            ['text' => 'Well done!', 'format' => FORMAT_HTML],
+        ];
+        $fromform->atomcount = ['0'];
 
-        $fromform->otherfeedback = array('text' => 'That is not right.', 'format' => FORMAT_HTML);
+        $fromform->otherfeedback = ['text' => 'That is not right.', 'format' => FORMAT_HTML];
         $fromform->atomcount_other = '1';
         $fromform->penalty = 0.3333333;